Changed to a single Agent, NO MORE TYPES It was an error, to have concrete types as we depend on a configuration file in csv. I removed the factory and the types (Entity,Person and Sensor) also the interface, we now access directly to the Agent Implementation.
