--- src/main/java/main/asw/agents/Sensor.java
@@ -1,24 +1,0 @@
-package main.asw.agents;
-
-import main.asw.util.Checker;
-
-/**
- * 
- * @author Sergio Faya Fernandez
- * 
- */
-public class Sensor extends AbstractAgent {
-	
-	public Sensor(String name,String email, String id)  {
-		super(name, email, id);
-		this.agentKind = SENSOR;
-	}
-	
-	@Override
-	protected boolean validateId(String nif) {
-		//Check repetition in database
-		Checker.isNull(nif);
-		Checker.isEmpty(nif);
-		return true;
-	}
-}
